+++ packages/inula-dev-tools/babel.config.js
@@ -17,7 +17,7 @@ module.exports = api => {
   const isTest = api.env('test');
   console.log('isTest', isTest);
 
-  const plugins = [['@babel/plugin-proposal-class-properties', { loose: false }]];
+  const plugins = [['@babel/plugin-transform-class-properties', { loose: false }]];
 
   if (process.env.NODE_ENV !== 'production') {
     plugins.push(['@babel/plugin-transform-react-jsx-source']);
@@ -30,9 +30,8 @@ module.exports = api => {
       [
         '@babel/preset-react',
         {
-          runtime: 'classic',
-          pragma: 'Inula.createElement',
-          pragmaFrag: 'Inula.Fragment',
+          runtime: 'automatic', // 新增
+          importSource: 'openinula', // 新增
         },
       ],
     ],