+++ packages/inula-request/CHANGELOG.md
@@ -1,5 +1,17 @@
 # 0.0.10 版本
 
+## 1.1.9
+
+### Patch Changes
+
+- 52b7f52: feat: inula-request add redirect support; tweak openinula peerDependencies version
+
+## 1.1.8
+
+### Patch Changes
+
+- 9030693: fix: 修复inula-request请求头重复添加Content—Type的问题
+
 ## 1.1.5
 
 ### Patch Changes